+++ src/utils/useAuth.tsx
@@ -1,4 +1,4 @@
-import { AuthContext } from "../context/AuthContext"
+import { AuthContext } from "../context/AuthContext";
 import { useContext } from "react";
 export function useAuth() {
   const context = useContext(AuthContext);
